feat : Add definition of convolution of measures
Add definition of convolution of measures on monoids.
Add commutativity of convolution if the monoid is commutative.
Add convolution of measures is finite if both measures are finite.

This PR is mainly to invite contributions to establish more of the
theory of convolutions of measures.

Regarding docstring & copyright: I needed to fill this in to make
Github happy, feel free to change.