refactor(Analysis/Calculus/FormalMultilinearSeries): generalize to additive monoids
While many of the lemmas cease to apply,
the definitions of sequences and their sums are still perfectly well-behaved in additive monoids;
no negation is needed.

We could use this to generalize `NormedSpace.exp` such that it works with `NNReal`,
but unless leanprover-community/mathlib#16554 or similar lands and provides a `DivisionSemiring.nnqsmul` field,
that would conflict with #8370.