feat(Algebra/Homology): the cylinder of a homological complex (#9483)
In this PR, the cylinder object of a homological complex is constructed. It is used to show that a functor which inverts homotopy equivalences sends homotopic maps to the same map. In the future, it shall be used in order to show that homotopic maps become equal in the derived category.