refactor(Probability/Distributions/Exponential): exponential distrubtion as special case of Gamma distribution (#9462)
Refactor most proofs of `Probability/Distributions/Exponential.lean` in terms of the proofs in `Probability/Distributions/Gamma.lean`, as the exponential distribution is a special case of the Gamma distribution.

Adds `stronglyMeasurable_gammaPdfReal` to `Probability/Distributions/Gamma` and [@measurability] tags to the necessary results in `Probability/Distributions/Gamma` and `Probability/Distributions/Exponential. ` 

Style : improve consistency in notation.