chore: adaptations for leanprover/lean4#3124 (#9500)
These are adaptations required for Leo's implementation of simprocs from leanprover/lean4#3124. 

This is a PR to `bump/v4.6.0`. Before it can be merged we need to merge 
- [x] https://github.com/leanprover/std4/pull/496
- [x] https://github.com/leanprover-community/aesop/pull/93
and then to update the lakefile so both dependencies point at the `bump/v4.6.0` branches.

I've updated a number of Mathlib tactics to adjust to the new type signatures in the simplifier, and these changes need to be reviewed carefully.



Co-authored-by: Scott Morrison <scott.morrison@gmail.com>