fix: check stderr again in cache (#9752)
This reverts #9434, which itself was a revert of #9419. We use `--no-progress-meter` to avoid the issue with progress bars causing a failure, but we still want to see error reports. The failure message is also made clearer to avoid the `uncaught exception:` issue showing up in [recent builds](https://github.com/leanprover-community/mathlib4/actions/runs/7524334618/job/20478989708#step:17:4161).



Co-authored-by: Mario Carneiro <di.gama@gmail.com>