refactor(LinearAlgebra/SesquilinearForm): Sesquilinear Maps (#9312)
Some of the concepts in `LinearAlgebra/SesquilinearForm` can be generalised from Sesquilinear Forms to Sesquilinear Maps with little or no impact on the rest of Mathlib. This PR makes those generalisations.

Further generalisations are likely possible, but the scope of this PR is to only consider changes which do not require non-trivial modifications to other parts of Mathlib, or other sections in `SesquilinearForm.lean`. Further changes can be considered in subsequent PRs if desired.



Co-authored-by: Christopher Hoskin <mans0954@users.noreply.github.com>