chore(Algebra/Subalgebra): de-instance `Subalgebra.module'`
This turns `Subalgebra.module'` into a regular instance, since we don't actually seem to use it anywhere. So it looks like a good candidate for making it a scoped instance only, since it does make things quite expensive.

I tried this a week ago on `Submodule` instead of `Subalgebra`, and the former is used in a few places. But this one can be disabled for free.