refactor: Delete `Algebra.GroupPower.Lemmas`  (#9411)
`Algebra.GroupPower.Lemmas` used to be a big bag of lemmas that made it there on the criterion that they needed "more imports". This was completely untrue, as all lemmas could be moved to earlier files in PRs:
- #9440
- #9442
- #9443
- #9455
- #9456
- #9457
- #9459
- #9461
- #9463
- #9466
- #9501
- #9502
- #9503
- #9505
- #9551
- #9553
- #9720
- #9739
- #9740
- #9805
- #9806
- and this one

There are several reasons for this:
* Necessary lemmas have been moved to earlier files since lemmas were dumped in `Algebra.GroupPower.Lemmas`
* In the Lean 3 → Lean 4 transition, Std acquired basic `Int` and `Nat` lemmas which let us shortcircuit the part of the algebraic order hierarchy on which the corresponding general lemmas rest
* Some proofs were overpowered
* Some earlier files were tangled and I have untangled them

This PR finishes the job by moving the last few lemmas out of `Algebra.GroupPower.Lemmas`, which is therefore deleted.