feat : Move `IsUniform`, refactor
Move `IsUniform` from `Probability/Density` to
`Probability/Distributions/Uniform` to make it more in line with the
other distributions.

Refactor `IsUniform` in terms of the new `IsUniformVolume`, which
abstracts away from the random variable and probability space.

This PR will attempt to also refactor
`Probability/ProbabilityMassFunction/Uniform` in such a way that this
follows naturally from `IsUniform`.