feat : step towards linking PMF.
The next step would be to show that our usual PDF agrees with the one
that we picked, and then to derive the PMF from there. I have added
two definitions and show that they agree with `pdf ...` and each-other.

I need to link this to PMF, but I'm stuck on getting this to work.
I probably won't have time to work on this the next days, so feel free
to pick this up.