feat: Flooring, ceiling roots of a natural number (#8839)
This PR defines the flooring and ceiling root of a natural number. The `n`-th flooring/ceiling root of `a` is the natural number whose `p`-adic valuation is the floor/ceil of the `p`-adic valuation of `a`. Those operations are respectively the right and left adjoints to the map `a ↦ a ^ n`.

For example the `2`-nd flooring and ceiling roots of `2^3 * 3^2 * 5` are `2 * 3` and `2^2 * 3 * 5` respectively.