Prioritize section on mathlib4 dependency usage
This is a step towards focusing post-installation steps in the README
better and organizing them in order of importance.

Installation of mathlib4 is never explicitly explained with the main
"experimenting" link focusing on installing Lean4 itself. As most of
the setup and making the library available happens when creating a Lean
workspace, put related sections closer to follow right after the fairly
short installation section.