chore: cleanup use of simp singlePass option (#9928)
Of the 18 uses of `singlePass`, in 3 cases we can just use `rw`, in 4 cases it isn't needed at all.

In the other 11 cases we are always use it as `simp (config := {singlePass := true}) only [X]` (i.e. with just a single simp lemma), and that `simp` call would loop forever (usually failing with a maxRecDepth error, sometimes with heartbeats). I've left these as is.

There's also one case where there was a missing `only`.



Co-authored-by: Scott Morrison <scott.morrison@gmail.com>