chore: cache: do not try to upload files just downloaded (#9687)
If `put` in CI finds that an archive already exists in the local cache, we can assume it exists in the cloud cache as well. Should provide a good-enough avoidance of repeatedly uploading the same archive on hosts like Cloudflare that do not support no-overwrite and still save a ton of write requests on Azure.

Co-authored-by: Mario Carneiro <di.gama@gmail.com>