fix: `shake --downstream` bug (#10469)
The old code was not taking into account a certain import situation, and the fix requires recalculating the whole transitive dependency graph, which is a bit expensive. Furthermore, the output is now more nuanced, because instead of saying "remove A, instead import A in B1, B2", it may need to suggest importing other files like "remove A, instead import C1 in B1 and C2, C3 in B2". It could suggest to import `A` in all affected downstream files, but those files might not be using `A` directly so this would cause shake to complain again after applying the fix.



Co-authored-by: Mario Carneiro <di.gama@gmail.com>