feat: the homotopy category of cochain complexes is pretriangulated (#9032)
This PR defines the pretriangulated structure on the homotopy category of cochain complexes in an additive category.

This result first appeared in the Liquid Tensor Experiment. In the LTE, the formalization followed the Stacks Project: in particular, the distinguished triangles were defined using degreewise-split short exact sequences of cochain complexes. Here, we follow the original definitions in Verdier's thesis (with the better sign conventions from the introduction of Brian Conrad's book *Grothendieck duality and base change*).



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>