style(Order): use `where __ := x` instead of `:= { x with }` (#10588)
The former style encourages one parent per line, which is easier to review and see diffs of. It also allows a handful of `:= fun a b =>`s to be replaced with the less noisy `a b :=`.

A small number of `inferInstanceAs _ with`s were removed, as they are automatic in those places anyway.