perf: speed up galActionHom_bijective_of_prime_degree (#10608)
This takes the proof from approximately 3 seconds to half a second on my laptop.

convert spent a fair amount of time dealing with the equality of Fintype.card calls with the same type but different Fintype instances. This sidesteps the issue by translating to Nat.card, which doesn't have the instance argument.