feat: `Pairwise` and composition with an equiv
Add a lemma about `Pairwise` being preserved by composing the
arguments of the pairwise relation with an equiv.

```lean
lemma EquivLike.pairwise_comp_iff {X : Type*} {Y : Type*} {F} [EquivLike F Y X]
    (f : F) (p : X → X → Prop) : Pairwise (fun y z ↦ p (f y) (f z)) ↔ Pairwise p := by
```

This depends on `Mathlib.Data.FunLike.Equiv` and
`Mathlib.Logic.Pairwise`, neither of which imports the other, so to
avoid increasing the imports of very basic files I put it in its own
new file, but feel free to suggest a better location of this lemma
(and to golf the proof).

From AperiodicMonotilesLean.