feat(CategoryTheory/GradedObject): more about the action of a composition of bifunctors (#10667)
The action on graded objects of a trifunctor obtained by composition of two bifunctors can be computed as a composition of the actions of these two bifunctors. In this PR, we consider the case when we apply a bifunctor to the second and third variables and then a bifunctor to the first variable and the result of this. The new code is very similar to #8242.