refactor(LinearAlgebra/QuadraticForm): Replace `BilinForm` with a scalar valued bi `LinearMap` (#10238)
Following on from #10097, which converted the companion of a quadratic form with a bilinear map, this PR replaces a number of results about quadratic forms and bilinear forms with results about quadratic forms and scalar valued bilinear maps. The long term aim is to be able to consider quadratic maps.

The main change is to `LinearAlgebra/QuadraticForm/Basic`, but this necessitates changes throughout `LinearAlgebra/QuadraticForm/`. Minor changes are also required elsewhere:

- `LinearAlgebra/CliffordAlgebra/`
- `LinearAlgebra/Matrix/PosDef`
- `LinearAlgebra/SesquilinearForm`
- A number of additional results about tensor products and linear maps are also required.



Co-authored-by: Eric Wieser <wieser.eric@gmail.com>
Co-authored-by: Christopher Hoskin <christopher.hoskin@overleaf.com>