feat(Logic/Equiv/Defs): composing an `EquivLike` with its `symm` (#10911)
If you convert an `EquivLike` to an `Equiv` in order to use `symm`, you can get combinations of the coercion of that `EquivLike` direct to a function, and the coercion of that `symm` to a function, which should simplify but `simp` fails to do anything with.  Add four `simp` lemmas to enable `simp` to act on such combinations.

From AperiodicMonotilesLean.